fix: Most likely fixed choppiness by not adding blocks during draw
The function `drawMap()' was re-adding block to world at every single
call to 'draw()`. That process was now split into two pieces. The map
is now read once and blocks are pushed to a table and `drawMap` now
doesn't even do nested for loops, it just iterates over the list of
blocks to draw.